# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "面向个人房东的租房管理后端系统，支持房源管理、租客管理、水电气收费、租金账单等功能。",
        "title": "租房管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/contracts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取合同列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "合同管理"
                ],
                "summary": "获取合同列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为租客签订租房合同",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "合同管理"
                ],
                "summary": "签订合同",
                "parameters": [
                    {
                        "description": "合同信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.CreateContractRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/contracts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取合同详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "合同管理"
                ],
                "summary": "获取合同详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "合同ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/contracts/{id}/terminate": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "终止合同",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "合同管理"
                ],
                "summary": "解约",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "合同ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "解约原因",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.TerminateContractRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/tenants": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取房东的租客列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租客管理"
                ],
                "summary": "获取租客列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新租客",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租客管理"
                ],
                "summary": "创建租客",
                "parameters": [
                    {
                        "description": "租客信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.CreateTenantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/tenants/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取租客详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租客管理"
                ],
                "summary": "获取租客详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "租客ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新租客信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租客管理"
                ],
                "summary": "更新租客",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "租客ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "租客信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.UpdateTenantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除租客",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租客管理"
                ],
                "summary": "删除租客",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "租客ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "rental-management_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rental-management_internal_service.CreateContractRequest": {
            "type": "object",
            "required": [
                "end_date",
                "monthly_rent",
                "room_id",
                "start_date",
                "tenant_id"
            ],
            "properties": {
                "deposit": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "monthly_rent": {
                    "type": "number"
                },
                "payment_day": {
                    "type": "integer"
                },
                "room_id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "integer"
                }
            }
        },
...
        "rental-management_internal_service.CreateTenantRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "emergency_contact": {
                    "type": "string"
                },
                "emergency_phone": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "id_card": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "rental-management_internal_service.TerminateContractRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "rental-management_internal_service.UpdateTenantRequest": {
            "type": "object",
            "properties": {
                "emergency_contact": {
                    "type": "string"
                },
                "emergency_phone": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "id_card": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}